+++ src/pages/profile-update/components/ProfileUpdateBaseForm.tsx
@@ -6,7 +6,7 @@ import { useForm } from 'react-hook-form'
 import { profileValidationRules } from '@src/utils/validation/authValidationRules'
 import { Button } from '@src/components/commons/button/Button'
 import { cn } from '@src/utils/cn'
-import type { MyPageData } from './my-page/components/ProfileData'
+import type { MyPageData } from '../../my-page/components/ProfileData'
 import { formatBirthDate } from '@src/utils/formatBirthDate'
 import { checkNickname } from '@src/api/auth'
 import { useState, useEffect } from 'react'
